Move e-mail operator to core
