+++ airflow/operators/email_operator.py
@@ -15,14 +15,14 @@
 # KIND, either express or implied.  See the License for the
 # specific language governing permissions and limitations
 # under the License.
-"""This module is deprecated. Please use `airflow.providers.email.operators.email`."""
+"""This module is deprecated. Please use `airflow.operators.email`."""
 
 import warnings
 
 # pylint: disable=unused-import
-from airflow.providers.email.operators.email import EmailOperator  # noqa
+from airflow.operators.email import EmailOperator  # noqa
 
 warnings.warn(
-    "This module is deprecated. Please use `airflow.providers.email.operators.email`.",
+    "This module is deprecated. Please use `airflow.operators.email`.",
     DeprecationWarning, stacklevel=2
 )
